feat(node): validate Piltover core contract on startup
Reads `get_program_info` from the on-chain core contract referenced by
`SettlementLayer::Starknet` at the top of `Node::launch` and aborts
startup if the values don't match the chain spec. Validation is
proof-mode aware: ZK chains check SNOS / layout-bridge / bootloader
program hashes plus the SNOS config hash; TEE chains check the SNOS
config hash only.

Adds a `proof_kind` field to `SettlementLayer::Starknet` (defaulting to
`ValidityProof` so existing chain spec files still parse) populated by
`katana init` from the selected proof implementation. Extracts
`SettlementChainProvider`, the program-hash constants, and the config
hash helper from the init module into a shared
`katana-sequencer-node::settlement_check` module so init's deploy-time
check and the new launch-time check share one implementation.
