+++ bin/katana/Cargo.toml
@@ -17,6 +17,7 @@ katana-primitives.workspace = true
 katana-provider.workspace = true
 katana-rpc-api.workspace = true
 katana-rpc-types.workspace = true
+katana-sequencer-node.workspace = true
 katana-starknet.workspace = true
 katana-utils.workspace = true
 
@@ -25,7 +26,6 @@ ratatui = "0.29"
 
 cairo-lang-starknet-classes.workspace = true
 anyhow.workspace = true
-async-trait.workspace = true
 byte-unit = "5.1.4"
 cainome.workspace = true
 clap.workspace = true
