+++ bin/katana/src/cli/init/deployment.rs
@@ -1,12 +1,15 @@
 use anyhow::{anyhow, Result};
 use cainome::cairo_serde;
+use katana_chain_spec::SettlementProofKind;
 use katana_contracts::piltover::Appchain;
 use katana_primitives::block::{BlockHash, BlockNumber};
-use katana_primitives::cairo::ShortString;
 use katana_primitives::class::ContractClass;
 use katana_primitives::{felt, ContractAddress, Felt};
 use katana_rpc_api::tee::compute_katana_tee_config_hash;
 use katana_rpc_types::class::Class;
+use katana_sequencer_node::settlement_check::{
+    self, SettlementChainProvider, SettlementValidationError,
+};
 use katana_starknet::rpc::StarknetRpcClient as StarknetClient;
 use katana_utils::{TxWaiter, TxWaitingError};
 use piltover::{
@@ -16,51 +19,17 @@ use piltover::{
 use spinoff::{spinners, Color, Spinner};
 use starknet::accounts::{Account, AccountError, ConnectedAccount, SingleOwnerAccount};
 use starknet::contract::{ContractFactory, UdcSelector};
-use starknet::core::crypto::compute_hash_on_elements;
 use starknet::core::types::{BlockId, BlockTag, FlattenedSierraClass, StarknetError};
 use starknet::providers::{Provider, ProviderError};
 use starknet::signers::LocalWallet;
+use settlement_check::{
+    compute_starknet_os_config_hash, BOOTLOADER_PROGRAM_HASH, LAYOUT_BRIDGE_PROGRAM_HASH,
+    SNOS_PROGRAM_HASH,
+};
 use thiserror::Error;
 use tracing::trace;
 
-use super::settlement::SettlementChainProvider;
-
 type SettlementInitializerAccount = SingleOwnerAccount<SettlementChainProvider, LocalWallet>;
-
-/// The StarknetOS program (SNOS) is the cairo program that executes the state
-/// transition of a new Katana block from the previous block.
-/// This program hash is required to be known by the settlement contract in order to
-/// only accept a new state from a valid SNOS program.
-///
-/// This program can be found here: <https://github.com/starkware-libs/sequencer/blob/v0.16.0-rc.0/crates/apollo_starknet_os_program/src/cairo/starkware/starknet/core/os/os.cairo>.
-const SNOS_PROGRAM_HASH: Felt =
-    felt!("0x10e5341a417427d140af8f5def7d2cc687d84591ff8ec241623c590b5ca8c80");
-
-/// To execute the SNOS program, a specific layout named "all_cairo" is required.
-/// However, this layout can't be verified by the Cairo verifier that lives on Starknet.
-///
-/// This is why we're using an other program, the Layout Bridge program, which act as a verifier
-/// written in Cairo which uses a layout supported by the Cairo verifier.
-///
-/// By verifying a SNOS proof using the Layout Bridge program, a new proof is generated which can be
-/// verified by the Cairo verifier.
-///
-/// For the same reason as above, the Layout Bridge program is required to be known by the
-/// settlement contract for security reasons.
-///
-/// This program can be found here: <https://github.com/starkware-libs/cairo-lang/blob/8276ac35830148a397e1143389f23253c8b80e93/src/starkware/cairo/cairo_verifier/layouts/all_cairo/cairo_verifier.cairo>.
-const LAYOUT_BRIDGE_PROGRAM_HASH: Felt =
-    felt!("0x43c5c4cc37c4614d2cf3a833379052c3a38cd18d688b617e2c720e8f941cb8");
-
-/// The bootloader program hash is the program hash of the bootloader program.
-///
-/// This program is used to run the layout bridge program in SHARP. This program hash is also
-/// required since the fact is computed based on the bootloader program hash and its output.
-///
-/// TODO: waiting for SHARP team to confirm if it's a custom bootloader or if we
-/// can find it in cairo-lang.
-const BOOTLOADER_PROGRAM_HASH: Felt =
-    felt!("0x5ab580b04e3532b6b18f81cfa654a05e29dd8e2352d88df1e765a84072db07");
 
 /// Default STRK fee-token address for chains generated by `katana init`.
 ///
@@ -277,19 +246,17 @@ pub async fn deploy_settlement_contract(
 
 /// Checks that the settlement contract is correctly configured.
 ///
-/// The values checked depend on the selected settlement mode (`tee`):
-/// * StarknetOs mode — bootloader / SNOS-program / layout-bridge program hashes and the
-///   `StarknetOsConfig3`-tagged config hash.
-/// * KatanaTee mode — the `KatanaTeeConfig1`-tagged environment config hash.
-///
-/// In both modes, the fact-registry address is compared against `expected_fact_registry`.
+/// Delegates the program-info comparison (program hashes for `StarknetOs`, config hash for
+/// `KatanaTee`, and cross-mode mismatch) to [`settlement_check::validate_starknet_settlement`] —
+/// the same code path that runs on every node startup. Then verifies the fact-registry address
+/// against `expected_fact_registry` (which is not checked at startup).
 ///
 /// `expected_fact_registry` is the Herodotus Atlantic address in StarknetOs mode and the
 /// `IAMDTeeRegistry` address in TEE mode. `tee` must match the variant the contract was
-/// constructed with; cross-mode lookup is reported as `InvalidProgramInfoVariant`.
+/// constructed with; cross-mode lookup surfaces as
+/// [`SettlementValidationError::InvalidProgramInfoVariant`] via the `Settlement` variant.
 ///
-/// On success, returns the SNOS config hash (the value that was verified to match the on-chain
-/// `snos_config_hash`).
+/// On success, returns the config hash that was verified.
 pub async fn check_program_info(
     chain_id: Felt,
     appchain_address: ContractAddress,
@@ -297,78 +264,21 @@ pub async fn check_program_info(
     expected_fact_registry: Felt,
     tee: bool,
 ) -> Result<Felt, ContractInitError> {
+    let proof_kind =
+        if tee { SettlementProofKind::Tee } else { SettlementProofKind::ValidityProof };
+
+    settlement_check::validate_starknet_settlement(
+        chain_id,
+        DEFAULT_FEE_TOKEN_ADDRESS,
+        appchain_address,
+        provider,
+        proof_kind,
+    )
+    .await?;
+
     let appchain = AppchainContractReader::new(appchain_address.into(), provider);
-
-    // Assert that the values are correctly set
-    let (program_info_res, facts_registry_res) =
-        tokio::join!(appchain.get_program_info().call(), appchain.get_facts_registry().call());
-
-    let actual_program_info = program_info_res.map_err(|e| ContractInitError::Other(anyhow!(e)))?;
-    let facts_registry = facts_registry_res.map_err(|e| ContractInitError::Other(anyhow!(e)))?;
-
-    let config_hash = match (tee, actual_program_info) {
-        (false, ProgramInfo::StarknetOs(info)) => {
-            let expected_config_hash = compute_config_hash(chain_id, DEFAULT_FEE_TOKEN_ADDRESS);
-
-            if info.layout_bridge_program_hash != LAYOUT_BRIDGE_PROGRAM_HASH {
-                return Err(ContractInitError::InvalidLayoutBridgeProgramHash {
-                    actual: info.layout_bridge_program_hash,
-                    expected: LAYOUT_BRIDGE_PROGRAM_HASH,
-                });
-            }
-
-            if info.bootloader_program_hash != BOOTLOADER_PROGRAM_HASH {
-                return Err(ContractInitError::InvalidBootloaderProgramHash {
-                    actual: info.bootloader_program_hash,
-                    expected: BOOTLOADER_PROGRAM_HASH,
-                });
-            }
-
-            if info.snos_program_hash != SNOS_PROGRAM_HASH {
-                return Err(ContractInitError::InvalidSnosProgramHash {
-                    actual: info.snos_program_hash,
-                    expected: SNOS_PROGRAM_HASH,
-                });
-            }
-
-            if info.snos_config_hash != expected_config_hash {
-                return Err(ContractInitError::InvalidConfigHash {
-                    actual: info.snos_config_hash,
-                    expected: expected_config_hash,
-                });
-            }
-
-            info.snos_config_hash
-        }
-
-        (true, ProgramInfo::KatanaTee(info)) => {
-            let expected_config_hash =
-                compute_katana_tee_config_hash(chain_id, DEFAULT_FEE_TOKEN_ADDRESS);
-
-            if info.katana_tee_config_hash != expected_config_hash {
-                return Err(ContractInitError::InvalidKatanaTeeConfigHash {
-                    actual: info.katana_tee_config_hash,
-                    expected: expected_config_hash,
-                });
-            }
-
-            info.katana_tee_config_hash
-        }
-
-        (false, ProgramInfo::KatanaTee(_)) => {
-            return Err(ContractInitError::InvalidProgramInfoVariant {
-                expected: "StarknetOs",
-                actual: "KatanaTee",
-            });
-        }
-
-        (true, ProgramInfo::StarknetOs(_)) => {
-            return Err(ContractInitError::InvalidProgramInfoVariant {
-                expected: "KatanaTee",
-                actual: "StarknetOs",
-            });
-        }
-    };
+    let facts_registry =
+        appchain.get_facts_registry().call().await.map_err(|e| ContractInitError::Other(anyhow!(e)))?;
 
     if facts_registry != expected_fact_registry.into() {
         return Err(ContractInitError::InvalidFactRegistry {
@@ -377,6 +287,11 @@ pub async fn check_program_info(
         });
     }
 
+    let config_hash = if tee {
+        compute_katana_tee_config_hash(chain_id, DEFAULT_FEE_TOKEN_ADDRESS)
+    } else {
+        compute_starknet_os_config_hash(chain_id, DEFAULT_FEE_TOKEN_ADDRESS)
+    };
     Ok(config_hash)
 }
 
@@ -397,7 +312,7 @@ fn build_program_info(tee: bool, chain_id: Felt) -> ProgramInfo {
         })
     } else {
         ProgramInfo::StarknetOs(StarknetOsProgramInfo {
-            snos_config_hash: compute_config_hash(chain_id, DEFAULT_FEE_TOKEN_ADDRESS),
+            snos_config_hash: compute_starknet_os_config_hash(chain_id, DEFAULT_FEE_TOKEN_ADDRESS),
             snos_program_hash: SNOS_PROGRAM_HASH,
             bootloader_program_hash: BOOTLOADER_PROGRAM_HASH,
             layout_bridge_program_hash: LAYOUT_BRIDGE_PROGRAM_HASH,
@@ -417,43 +332,11 @@ pub enum ContractInitError {
     #[error("failed to initialize contract: {0:#?}")]
     Initialization(AccountError<<SettlementInitializerAccount as Account>::SignError>),
 
-    #[error(
-        "invalid program info: layout bridge program hash mismatch - expected {expected:#x}, got \
-         {actual:#x}"
-    )]
-    InvalidLayoutBridgeProgramHash { expected: Felt, actual: Felt },
-
-    #[error(
-        "invalid program info: bootloader program hash mismatch - expected {expected:#x}, got \
-         {actual:#x}"
-    )]
-    InvalidBootloaderProgramHash { expected: Felt, actual: Felt },
-
-    #[error(
-        "invalid program info: snos program hash mismatch - expected {expected:#x}, got \
-         {actual:#x}"
-    )]
-    InvalidSnosProgramHash { expected: Felt, actual: Felt },
-
-    #[error(
-        "invalid program info: config hash mismatch - expected {expected:#x}, got {actual:#x}"
-    )]
-    InvalidConfigHash { expected: Felt, actual: Felt },
-
-    #[error(
-        "invalid program info: katana tee config hash mismatch - expected {expected:#x}, got \
-         {actual:#x}"
-    )]
-    InvalidKatanaTeeConfigHash { expected: Felt, actual: Felt },
-
-    #[error(
-        "invalid program info: settlement-mode mismatch - expected {expected} variant, got \
-         {actual}"
-    )]
-    InvalidProgramInfoVariant { expected: &'static str, actual: &'static str },
-
     #[error("invalid program state: fact registry mismatch - expected {expected:}, got {actual}")]
     InvalidFactRegistry { expected: Felt, actual: Felt },
+
+    #[error(transparent)]
+    Settlement(#[from] SettlementValidationError),
 
     #[error(transparent)]
     TxWaitingError(#[from] TxWaitingError),
@@ -480,38 +363,3 @@ fn prepare_contract_declaration_params(class: ContractClass) -> Result<Flattened
     Ok(class.try_into()?)
 }
 
-// NOTE: The reason why we're using the same address for both fee tokens is because we don't yet
-// support having native fee token on the chain.
-fn compute_config_hash(chain_id: Felt, fee_token: Felt) -> Felt {
-    compute_starknet_os_config_hash(chain_id, fee_token)
-}
-
-// https://github.com/starkware-libs/sequencer/blob/e13acc4c582352e777f5beae3476d157e6bdf4cf/crates/apollo_starknet_os_program/src/cairo/starkware/starknet/core/os/os_config/os_config.cairo#L10
-fn compute_starknet_os_config_hash(chain_id: Felt, fee_token: Felt) -> Felt {
-    // A constant representing the StarkNet OS config version.
-    const STARKNET_OS_CONFIG_VERSION: ShortString = ShortString::from_ascii("StarknetOsConfig3");
-    compute_hash_on_elements(&[STARKNET_OS_CONFIG_VERSION.into(), chain_id, fee_token])
-}
-
-#[cfg(test)]
-mod tests {
-    use katana_primitives::{felt, Felt};
-    use starknet::core::chain_id::{MAINNET, SEPOLIA};
-
-    use super::compute_starknet_os_config_hash;
-
-    const STRK_FEE_TOKEN: Felt =
-        felt!("0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d");
-
-    // Source:
-    //
-    // - https://github.com/starkware-libs/cairo-lang/blob/v0.14.0.1/src/starkware/starknet/core/os/os_config/os_config_hash.json
-    // - https://docs.starknet.io/tools/important-addresses/#fee_tokens
-    #[rstest::rstest]
-    #[case::mainnet(felt!("0x70c7b342f93155315d1cb2da7a4e13a3c2430f51fb5696c1b224c3da5508dfb"), MAINNET)]
-    #[case::testnet(felt!("0x1b9900f77ff5923183a7795fcfbb54ed76917bc1ddd4160cc77fa96e36cf8c5"), SEPOLIA)]
-    fn calculate_config_hash(#[case] config_hash: Felt, #[case] chain: Felt) {
-        let computed = compute_starknet_os_config_hash(chain, STRK_FEE_TOKEN);
-        assert_eq!(computed, config_hash);
-    }
-}
