+++ crates/chain-spec/tests/rollup.rs
@@ -46,6 +46,7 @@ fn chain_spec(n_dev_accounts: u16, with_balance: bool) -> ChainSpec {
         id: ChainId::default(),
         core_contract: Default::default(),
         rpc_url: Url::parse("http://localhost:5050").unwrap(),
+        proof_kind: Default::default(),
     };
 
     ChainSpec { id, genesis, settlement, fee_contracts }
