+++ crates/node/sequencer/Cargo.toml
@@ -29,11 +29,14 @@ katana-tasks.workspace = true
 katana-tee.workspace = true
 
 anyhow.workspace = true
+async-trait.workspace = true
 futures.workspace = true
 http.workspace = true
 jsonrpsee.workspace = true
 num-traits.workspace = true
+piltover.workspace = true
 starknet.workspace = true
+thiserror.workspace = true
 tower.workspace = true
 tracing.workspace = true
 url.workspace = true
@@ -49,5 +52,6 @@ tee-mock = [ "katana-tee/tee-mock" ]
 
 [dev-dependencies]
 katana-contracts.workspace = true
+rstest.workspace = true
 tempfile.workspace = true
 tokio = { workspace = true, features = [ "macros", "rt-multi-thread" ] }
